Simplifying function calls, fix file identation, added real_entry_list
From now on the view class owns the real_entry_list which holds all
visible entries in the view with direct relation. So the first index
stands for the entry of the line 1 and so on. Nevertheless theres an
offset between real_entry_list index and line filetab line number caused
by:
- Lines start in 0 and array in 1
- Header size above the entries there's a header
So real_entry_list[1] isn't on the first line.
This is acconted by real_entry_list[line_number - INFO.HEADER_SIZE + 1]